+++ packages/wonder-blocks-accordion/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @khanacademy/wonder-blocks-accordion
 
+## 3.1.16
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-typography@4.0.0
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+    - @khanacademy/wonder-blocks-clickable@7.1.4
+    - @khanacademy/wonder-blocks-core@12.3.0
+    - @khanacademy/wonder-blocks-icon@5.1.4
+
+## 3.1.15
+
+### Patch Changes
+
+- Updated dependencies [1338494]
+    - @khanacademy/wonder-blocks-tokens@10.2.1
+    - @khanacademy/wonder-blocks-clickable@7.1.3
+    - @khanacademy/wonder-blocks-typography@3.2.4
+
 ## 3.1.14
 
 ### Patch Changes