+++ packages/wonder-blocks-birthday-picker/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @khanacademy/wonder-blocks-birthday-picker
 
+## 4.0.8
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-typography@4.0.0
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+    - @khanacademy/wonder-blocks-core@12.3.0
+    - @khanacademy/wonder-blocks-dropdown@10.1.3
+    - @khanacademy/wonder-blocks-icon@5.1.4
+    - @khanacademy/wonder-blocks-layout@3.1.15
+
+## 4.0.7
+
+### Patch Changes
+
+- Updated dependencies [1338494]
+    - @khanacademy/wonder-blocks-tokens@10.2.1
+    - @khanacademy/wonder-blocks-dropdown@10.1.2
+    - @khanacademy/wonder-blocks-layout@3.1.14
+    - @khanacademy/wonder-blocks-typography@3.2.4
+
 ## 4.0.6
 
 ### Patch Changes