+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 7.1.4
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+    - @khanacademy/wonder-blocks-core@12.3.0
+
+## 7.1.3
+
+### Patch Changes
+
+- Updated dependencies [1338494]
+    - @khanacademy/wonder-blocks-tokens@10.2.1
+
 ## 7.1.2
 
 ### Patch Changes