+++ packages/wonder-blocks-icon-button/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @khanacademy/wonder-blocks-icon-button
 
+## 10.1.1
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+    - @khanacademy/wonder-blocks-clickable@7.1.4
+    - @khanacademy/wonder-blocks-core@12.3.0
+    - @khanacademy/wonder-blocks-icon@5.1.4
+    - @khanacademy/wonder-blocks-styles@0.2.11
+    - @khanacademy/wonder-blocks-theming@3.4.0
+
+## 10.1.0
+
+### Minor Changes
+
+- 1338494: Adds the `thunderblocks` theme to the IconButton component.
+
+### Patch Changes
+
+- 2779e43: Refactor `IconButton` internally to create an `IconButtonUnstyled` component that can be reused for more button components. Fixed visual styles with keyboard navigation."
+- Updated dependencies [1338494]
+    - @khanacademy/wonder-blocks-tokens@10.2.1
+    - @khanacademy/wonder-blocks-clickable@7.1.3
+    - @khanacademy/wonder-blocks-styles@0.2.10
+
 ## 10.0.0
 
 ### Major Changes