+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 5.1.18
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-typography@4.0.0
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+    - @khanacademy/wonder-blocks-core@12.3.0
+    - @khanacademy/wonder-blocks-form@7.1.16
+    - @khanacademy/wonder-blocks-icon@5.1.4
+    - @khanacademy/wonder-blocks-icon-button@10.1.1
+
+## 5.1.17
+
+### Patch Changes
+
+- Updated dependencies [1338494]
+- Updated dependencies [2779e43]
+- Updated dependencies [1338494]
+    - @khanacademy/wonder-blocks-tokens@10.2.1
+    - @khanacademy/wonder-blocks-icon-button@10.1.0
+    - @khanacademy/wonder-blocks-form@7.1.15
+    - @khanacademy/wonder-blocks-typography@3.2.4
+
 ## 5.1.16
 
 ### Patch Changes