+++ packages/wonder-blocks-styles/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-styles
 
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [1338494]
+    - @khanacademy/wonder-blocks-tokens@10.2.1
+
 ## 0.2.9
 
 ### Patch Changes