+++ packages/wonder-blocks-typography/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @khanacademy/wonder-blocks-typography
 
+## 4.0.0
+
+### Major Changes
+
+- dd76e7c: - Add new Heading and BodyText Typography components with themable font and other styles.
+    - Change exported `Heading` and `BodyText` types for Typography component collections to free those names up for new components.
+
+### Patch Changes
+
+- Updated dependencies [dd76e7c]
+    - @khanacademy/wonder-blocks-tokens@10.3.0
+    - @khanacademy/wonder-blocks-core@12.3.0
+
+## 3.2.4
+
+### Patch Changes
+
+- Updated dependencies [1338494]
+    - @khanacademy/wonder-blocks-tokens@10.2.1
+
 ## 3.2.3
 
 ### Patch Changes